+++ www/components/form/admin-form/index.tsx
@@ -1,0 +1,7 @@
+import { AdminInput } from "./input";
+import { AdminFormRoot } from "./root";
+
+export default {
+  Input: AdminInput,
+  Root: AdminFormRoot,
+};
