--- www/components/form/index.tsx
@@ -1,7 +1,0 @@
-import { Input } from "./input";
-import { FormRoot } from "./root";
-
-export default {
-  Root: FormRoot,
-  Input,
-};
