+++ www/components/form/input.ts
@@ -1,8 +1,5 @@
 import type { DetailedHTMLProps, InputHTMLAttributes } from "react";
 
-import { AdminInput } from "./admin-input";
-import { PublicInput } from "./public-form/input";
-
 export type InputProps = DetailedHTMLProps<
   InputHTMLAttributes<HTMLInputElement>,
   HTMLInputElement
@@ -12,7 +9,3 @@ export type InputProps = DetailedHTMLProps<
   containerClassName?: string;
   admin?: boolean;
 };
-
-export function Input({ admin, ...props }: InputProps) {
-  return admin ? <AdminInput {...props} /> : <PublicInput {...props} />;
-}
